requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cffi==1.15.1
cryptography==41.0.1
importlib-metadata==6.7.0
jaraco.classes==3.2.3
keyring==24.0.1
more-itertools==9.1.0
mpmath==1.3.0
passlib==1.7.4
pycparser==2.21
pycryptodome==3.18.0
pyodbc==4.0.39
python-dotenv==1.0.0
pywin32-ctypes==0.2.1
sympy==1.12
zipp==3.15.0
